pdf_tools/requirements.txt
PyPDF2>=2.0.0
Pillow>=9.0.0
PyMuPDF>=1.18.0
ttkthemes>=3.2.0
